create and point favorites and check valid ticken

+++ controllers/favoritesController.ts
@@ -3,23 +3,21 @@ import { favoriteUserProduct } from '../models'
 
 //======================================
 // interfaces
-interface IRequest extends Request{
-    params:{
-        userId: string
-    }
-}
-
-interface IRequestAddOrRemoveProductOfFavorites extends IRequest{
+interface IRequestAddOrRemoveProductOfFavorites extends Request{
     body:{
         productId: number
-    }
+    },
 }
 
 //======================================
 // class-controller
 export default class FavoritesController{
     public static addProductToFavorites = async(req: IRequestAddOrRemoveProductOfFavorites, res: Response) => {
-        const userId = Number(req.params.userId)
+        if(!req.user){
+            return res.json({message: 'user is unlogin'})
+        }
+        const userId = req.user.id
+
         const [item, created] = await favoriteUserProduct.findOrCreate({
             where: { productId: req.body.productId, userId },
             defaults: {
@@ -37,7 +35,10 @@ export default class FavoritesController{
     }
 
     public static removeProductFromFavorites = async(req: IRequestAddOrRemoveProductOfFavorites, res: Response)=>{
-        const userId = Number(req.params.userId)
+        if(!req.user){
+            return res.json({message: 'user is unlogin'})
+        }
+        const userId = req.user.id
 
         await favoriteUserProduct.destroy({where:{
             userId,
